Safari-609.1.11.1/LayoutTests/css3/shapes/shape-outside/values/shape-outside-ellipse-007-expected.txt


FAIL ellipse(foo) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(10) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(20px 30px 40px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(20% 30% 40%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(20px, 30px, 40px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(20%, 30%, 40%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(20px 30px closest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(20px 30px farthest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(closest-side 20px 30px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(farthest-side 20px 30px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(20% 30% closest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(20% 30% farthest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(closest-side 20% 30%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(farthest-side 20% 30%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(closest-side farthest-side 20px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(closest-side farthest-side 20%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(farthest-side closest-side closest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL ellipse(farthest-side, closest-side, closest-side) - invalid assert_equals: expected (object) null but got (string) ""
